feat: Add public key storage in local storage
This commit adds the functionality to store the public key in the local storage when a user logs in. It modifies the `Login` component to save the public key in the local storage using the user's ID as the key. This change enables the client-side encryption feature by allowing the public key to be easily accessed when encrypting messages.